feat(script): turn in up to 10
players should now be able to turn in 1 or up to 10 sets at once.
Chossing the "Up to 10" option, will result in 10 checks if 1 item is available and a mail sent for each item. 
If 10 exchanges are succesful, 10 mails are sent. If less than 10 exchanges are succesful, a single error message will fire. But the player will still receive the succesful attempts via mail.